backend/requirements.txt
asgiref==3.8.1
Django==4.2.9
djangorestframework
gunicorn==22.0.0
psycopg2-binary
djangorestframework-simplejwt